Changed evaluation to a much simpler procedure.
First of all, these precision/recall estimates are done only on the positions of loci,
they completely ignore the cells which are being called or not.

Recall:
SVs are then being counted as "called" whenever there are a set of SV calls that
cover it to at least 80% (SV calls need not be in the same cells).

Precision:
Calls are considered to be true whenever they overlap SVs by at least one base pair.